wm: Try to clean up management of X vs. composited windows.
Users of the Window class have had to manage getting both
the composited and X (a.k.a. "client") windows to the
correct position, hiding the X window offscreen to disable
mouse input, etc.  This adds a new set of methods to
the Window class to set its visibility ("show it onscreen",
"show it onscreen but don't let it receive mouse events",
and "neither show it nor let it receive events") and a
Move() method that can then be used to move both the
composited and X window appropriately.

I'm not porting any of the existing code over to the new
methods in this change, but I may update some of it in
another change.

This change also fixes a few things that resulted in
compilation errors when I tried to build outside of the
chroot on an Ubuntu system recently.

BUG=none
TEST=added; also built an image, used it for a bit, and didn't notice anything broken (this is intended to be a no-op change)

Review URL: http://codereview.chromium.org/6646030

+++ compositor/real_compositor.cc
@@ -913,9 +913,11 @@ void RealCompositor::EnableDrawTimeout() {
         !last_draw_time_.is_null() ?
             GetMonotonicTime() - last_draw_time_ :
             TimeDelta();
-    int ms_until_draw =
-        max(FLAGS_draw_timeout_ms - time_since_draw.InMilliseconds(), 0LL);
-    event_loop_->ResetTimeout(draw_timeout_id_, ms_until_draw,
+    int64_t ms_until_draw =
+        max(FLAGS_draw_timeout_ms - time_since_draw.InMilliseconds(),
+            static_cast<int64_t>(0));
+    event_loop_->ResetTimeout(draw_timeout_id_,
+                              ms_until_draw,
                               FLAGS_draw_timeout_ms);
     draw_timeout_enabled_ = true;
   }
